+++ settings.php
@@ -3,7 +3,7 @@
     ini_set('display_errors', 0);
     error_reporting(E_ALL ^ E_NOTICE);
     
-    define('SLYR_WC_version',       "2.4");
+    define('SLYR_WC_version',       "2.5");
 
     global $wp_version;
     if (version_compare($wp_version,'4.5','>=')) {
@@ -47,7 +47,7 @@
     // Avoids wordpress to ask for credentials when testing on localhost
     if (!defined('FS_METHOD')) define('FS_METHOD',                 'direct');
 
-    define('SLYR_WC_DEBBUG',                0);
+    define('SLYR_WC_DEBBUG',                3); //Modificar a 0 al terminar
 
     if (!defined('SLYR_WC__PLUGIN_DIR')) define('SLYR_WC__PLUGIN_DIR', plugin_dir_path(__FILE__));
     if (!defined('SLYR_WC__LOGS_DIR')) define('SLYR_WC__LOGS_DIR', SLYR_WC__PLUGIN_DIR.'/logs/');